+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed AIFS-ENS invariants from model input / outputs and moved into model wrapper
 - CDS and NCAR lexica have been updated with variables for AIFS-ENS
+- Updated AIFS model to 1.1 checkpoint version
 
 ### Deprecated